Read contact numbers from Ramulator HDF5 files
In the HDF5 file, this is called "ports" because originally it was
not guaranteed that ENS port numbers and jacksheet numbers would
be the same. In practice, we always made sure this was the case and
in recent versions of Ramulator, this is guaranteed to be the case
when using bptools to generate the config file.